feat(htmlencoder.php): move to core/rendering
Cleanup architecture

Issue #17
